% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_tilda_addin.R
\name{insert_tilda_addin}
\alias{insert_tilda_addin}
\title{Insert " ~ ".
Call this function as an addin to insert \code{" ~ "} at the cursor
  position.}
\usage{
insert_tilda_addin()
}
\description{
Insert " ~ ".
Call this function as an addin to insert \code{" ~ "} at the cursor
  position.
}
