% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst.packs.R
\name{inst.packs}
\alias{inst.packs}
\title{Multiple package installation (without loading)}
\usage{
inst.packs(packlist = c(), dependencies = TRUE)
}
\arguments{
\item{packlist}{character vector of the names of packages whose current versions should be downloaded from the repositories.}

\item{dependencies}{logical indicating whether to also install uninstalled packages which these packages depend on/link to/import/suggest.}
}
\description{
The \code{inst.packs} function takes a single command, a list of package names, and will install them, if not already installed.
}
\examples{
inst.packs(c("ggplot2", "dplyr"))
}
\keyword{install}
\keyword{pack}
\keyword{package}
