% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packs.R
\name{install_packs}
\alias{install_packs}
\title{Multiple package installation (without loading)
The \code{install_packs} function takes a single command, a list of package
  names, and will install them, if not already installed.}
\usage{
install_packs(..., dependencies = TRUE, force = FALSE)
}
\arguments{
\item{...}{one or more character strings of package names
to be downloaded from the repositories}

\item{dependencies}{a boolean; if \code{TRUE} dependency packages will also
be installed, if \code{FALSE} dependency packages will not be installed}

\item{force}{logical indicating whether packages should be forced to be
installed/updated}
}
\description{
Multiple package installation (without loading)
The \code{install_packs} function takes a single command, a list of package
  names, and will install them, if not already installed.
}
\keyword{install}
\keyword{pack}
\keyword{package}
