% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\title{Extract the \code{n} characters from the left of a string}
\usage{
left(x, n)
}
\arguments{
\item{x}{a string}

\item{n}{the number of characters from the left of the string to extract}
}
\description{
\code{left} uses a specific case of the \code{substr} function to take only
  the left most \code{n} characters from the input string.
}
\examples{
foo <- "foo"
left(foo, 2)

}
