% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{\%out\%}
\alias{\%out\%}
\title{Value un-Matching}
\usage{
lhs \%out\% rhs
}
\arguments{
\item{lhs}{A value or vector to be un-matched from.}

\item{rhs}{A value or vector to be un-matched to.}
}
\description{
The \code{\%out\%} function works as the inverse of
  \code{\%in\%}, looking for the absence of value or values of the
  \code{lhs} from the \code{rhs}.
}
\examples{
v <- c("a","b","c")
u <- c("a","b","d")
v \%out\% u
}
\keyword{nin}
\keyword{notin}
\keyword{out}
