% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{Extract the \code{n} characters from the right of a string
\code{right} uses a specific case of the \code{substr} function to take only
  the right most \code{n} characters from the input string.}
\usage{
right(x, n)
}
\arguments{
\item{x}{a string}

\item{n}{the number of characters from the right of the string to extract}
}
\description{
Extract the \code{n} characters from the right of a string
\code{right} uses a specific case of the \code{substr} function to take only
  the right most \code{n} characters from the input string.
}
\examples{
foo <- "foo"
right(foo, 2)
}
