% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st.err.R
\name{st.err}
\alias{st.err}
\title{Calculate the standard error}
\usage{
st.err(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector for which the standard error will be calculated
from.}

\item{na.rm}{If *NA* values are not an issue, then *FALSE*,
otherwise, *TRUE*.}
}
\description{
The \code{st.err} function calculates the standard error of a variable,
taking into account \code{NA} values, either to keep or remove them
depending on the user input to the \code{na.rm} argument.
}
\examples{
x <- rnorm(10, 5, 0.25)
st.err(x)
}
\keyword{error}
\keyword{se}
\keyword{standard}
