% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.pack.R
\name{use.pack}
\alias{use.pack}
\title{Package installing and loading function}
\usage{
use.pack(package)
}
\arguments{
\item{package}{A package name as a quoted string.}
}
\description{
The use.pack function takes a single command, the name of a package, and will install it, if not already installed, and then load it.
}
\examples{
use.pack("ggplot2")
}
\keyword{install}
\keyword{pack}
\keyword{package}
\keyword{use}
