% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.packs.R
\name{use.packs}
\alias{use.packs}
\title{Package (multiple) installing and loading function}
\usage{
use.packs(packlist = c())
}
\arguments{
\item{packlist}{A list of package names as a vector of quoted strings.}
}
\description{
The use.packs function takes a single command, a list of package names, and will install them, if not already installed, and then load them.
}
\examples{
use.packs(c("ggplot2", "dplyr"))
}
\keyword{install}
\keyword{pack}
\keyword{package}
\keyword{use}
