% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pack.R
\name{use_pack}
\alias{use_pack}
\title{Package installing and loading function
\code{use_pack} takes a single command, the name of a package, and will
  install it, if not already installed, and then load it.}
\usage{
use_pack(package = NULL, dependencies = TRUE, force = FALSE)
}
\arguments{
\item{package}{a string; the name of a package to be installed}

\item{dependencies}{a boolean; if \code{TRUE} dependency packages will also
be installed, if \code{FALSE} dependency packages will not be installed}

\item{force}{logical indicating whether packages should be forced to be
installed/updated}
}
\description{
Package installing and loading function
\code{use_pack} takes a single command, the name of a package, and will
  install it, if not already installed, and then load it.
}
\keyword{install}
\keyword{pack}
\keyword{package}
\keyword{use}
