% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_packs.R
\name{use_packs}
\alias{use_packs}
\title{Package (multiple) installing and loading function
\code{use_packs} takes a single command, a list of package names, and will
  install them, if not already installed, and then load them.}
\usage{
use_packs(packlist = c(), dependencies = TRUE)
}
\arguments{
\item{packlist}{character vector of the names of packages whose current
versions should be downloaded from the repositories.}

\item{dependencies}{logical indicating whether to also install uninstalled
packages which these packages depend on/link to/import/suggest.}
}
\description{
Package (multiple) installing and loading function
\code{use_packs} takes a single command, a list of package names, and will
  install them, if not already installed, and then load them.
}
\keyword{install}
\keyword{pack}
\keyword{package}
\keyword{use}
