% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst.packs.R
\name{inst.packs}
\alias{inst.packs}
\title{Multiple package installation (without loading)}
\usage{
inst.packs(packlist = c())
}
\arguments{
\item{packlist}{A list of package names as a vector of quoted strings.}
}
\description{
The \code{inst.packs} function takes a single command, a list of package names, and will install them, if not already installed.
}
\examples{
inst.packs(c("ggplot2", "dplyr"))
}
\keyword{install}
\keyword{pack}
\keyword{package}
