% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst_packs.R
\name{inst_packs}
\alias{inst_packs}
\title{Multiple package installation (without loading)
The \code{inst_packs} function takes a single command, a list of package
  names, and will install them, if not already installed.}
\usage{
inst_packs(..., dependencies = TRUE)
}
\arguments{
\item{dependencies}{logical indicating whether to also install uninstalled
packages which these packages depend on/link to/import/suggest.}

\item{packlist}{character vector of the names of packages whose current
versions should be downloaded from the repositories.}
}
\description{
Multiple package installation (without loading)
The \code{inst_packs} function takes a single command, a list of package
  names, and will install them, if not already installed.
}
\keyword{install}
\keyword{pack}
\keyword{package}
