% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst_packs.R
\name{inst_packs}
\alias{inst_packs}
\title{Multiple package installation (without loading)
The \code{inst_packs} function takes a single command, a list of package
  names, and will install them, if not already installed.}
\usage{
inst_packs(..., dependencies = TRUE)
}
\arguments{
\item{...}{one or more character strings of package names
to be downloaded from the repositories}

\item{dependencies}{a boolean; if \code{TRUE} dependency packages will also
be installed, if \code{FALSE} dependency packages will not be installed}
}
\description{
Multiple package installation (without loading)
The \code{inst_packs} function takes a single command, a list of package
  names, and will install them, if not already installed.
}
\keyword{install}
\keyword{pack}
\keyword{package}
