% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.plot.R
\name{multi.plot}
\alias{multi.plot}
\title{Plotting of multiple objects in a single window}
\usage{
multi.plot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{further arguments (plot names) passed to or from other methods.}

\item{plotlist}{A vector list of plot objects, e.g. c(plot1, plot2, ..., plot_n).}

\item{cols}{Number of columns in layout.}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
The \code{multi.plot} function is meant to make plotting of multiple plot objects easier and more straightforward. Adapted from Winston Chang's "Cookbook for R".
}
\keyword{multi}
\keyword{plot}
