% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.pack.R
\name{use.pack}
\alias{use.pack}
\title{Package installing and loading function}
\usage{
use.pack(pack, dependencies = TRUE)
}
\arguments{
\item{pack}{character value of the name of a package whose current versions should be downloaded from the repositories.}

\item{dependencies}{logical indicating whether to also install uninstalled packages which this packages depend on/link to/import/suggest.}
}
\description{
The use.pack function takes a single command, the name of a package, and will install it, if not already installed, and then load it.
}
\examples{
use.pack("dplyr")
}
\keyword{install}
\keyword{pack}
\keyword{package}
\keyword{use}
