% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.packs.R
\name{use.packs}
\alias{use.packs}
\title{Package (multiple) installing and loading function}
\usage{
use.packs(packlist = c(), dependencies = TRUE)
}
\arguments{
\item{packlist}{character vector of the names of packages whose current versions should be downloaded from the repositories.}

\item{dependencies}{logical indicating whether to also install uninstalled packages which these packages depend on/link to/import/suggest.}
}
\description{
The use.packs function takes a single command, a list of package names, and will install them, if not already installed, and then load them.
}
\examples{
use.packs(c("ggplot2", "dplyr"))
}
\keyword{install}
\keyword{pack}
\keyword{package}
\keyword{use}
