% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pack.R
\name{use_pack}
\alias{use_pack}
\title{Package installing and loading function
\code{use_pack} takes a single command, the name of a package, and will
  install it, if not already installed, and then load it.}
\usage{
use_pack(pack, dependencies = TRUE)
}
\arguments{
\item{pack}{character value of the name of a package whose current versions
should be downloaded from the repositories.}

\item{dependencies}{logical indicating whether to also install uninstalled
packages which this packages depend on/link to/import/suggest.}
}
\description{
Package installing and loading function
\code{use_pack} takes a single command, the name of a package, and will
  install it, if not already installed, and then load it.
}
\keyword{install}
\keyword{pack}
\keyword{package}
\keyword{use}
