% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_packs.R
\name{use_packs}
\alias{use_packs}
\title{Package (multiple) installing and loading function
\code{use_packs} takes a single command, a list of package names, and will
  install them, if not already installed, and then load them.}
\usage{
use_packs(..., update = TRUE, dependencies = TRUE, force = FALSE)
}
\arguments{
\item{...}{one or more character strings of package names
to be downloaded from the repositories}

\item{update}{a boolean; if \code{TRUE} then any available updates for the
package will be installed, if \code{FALSE} then updates will not be
installed}

\item{dependencies}{a boolean; if \code{TRUE} dependency packages will also
be installed, if \code{FALSE} dependency packages will not be installed}

\item{force}{logical indicating whether packages should be forced to be
installed/updated}
}
\description{
Package (multiple) installing and loading function
\code{use_packs} takes a single command, a list of package names, and will
  install them, if not already installed, and then load them.
}
\keyword{install}
\keyword{pack}
\keyword{package}
\keyword{use}
